test(refactor): separate send lightning limits tests
